chore: Add env validate plugin for webpack
